+++ src/hooks/reactQuery/productOption/useProductOptionAllQuery.ts
@@ -1,18 +1,7 @@
-import productApi from '@/apis/product';
 import productOptionApi from '@/apis/productOption';
 import { useQuery } from '@tanstack/react-query';
 
-const useProductAll = () => {
-  const {
-    data: productAllResponse,
-    isLoading: isLoadingProducts,
-    error: productsError,
-  } = useQuery({
-    queryKey: ['getProductAll'],
-    queryFn: productApi.getProductAll,
-  });
-  const productAll = productAllResponse?.data;
-
+const useProductOptionAll = () => {
   const {
     data: optionAllResponse,
     isLoading: isLoadingOptions,
@@ -21,18 +10,16 @@ const useProductAll = () => {
     queryKey: ['getProductOptionAll'],
     queryFn: productOptionApi.getProductOptionAll,
   });
+
   const optionAll = optionAllResponse
     ? optionAllResponse.data.productOptions
     : [];
 
   return {
-    productAll,
     optionAll,
-    isLoadingProducts,
     isLoadingOptions,
-    productsError,
     optionsError,
   };
 };
 
-export default useProductAll;
+export default useProductOptionAll;
